blockmanager: Decouple block database code.
This decouples the code releated to loading and dumping the block
database from the block manager and moves it into a separate file.

This is part of an overall effort to remove all dependencies on the
global config variable from the block manager.
